alternative verification: no need to check r range
(not a serious pull request, just an observation for discussion)

I just noticed this while coding up a python-only Schnorr verifier -- the `r` from the signature never actually needs to be converted to an integer, it can stay as a bytes object the whole way through. It's not needed to check the range of `r` since the final step of comparing it against `x(R)` does this implicitly.

This only saves a negligible amount of CPU cycles, of course.